Use SegmentedKeyValueStorageAdapter as base class for Snappable adapter
